requirements.txt
streamlit==1.42.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.34
langchain-groq==0.2.1
python-dotenv==1.0.1
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0
Pillow==11.1.0
matplotlib==3.10.0
groq==0.13.1
ultralytics==8.3.82
